requirements.txt
openai==0.28
mysql-connector-python
python-dotenv
streamlit
faker
